requirements.txt
aiohttp==3.2.1
aioredis==1.2.0
alabaster==0.7.12
amqp==1.4.9
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
anyjson==0.3.3
arrow==0.14.2
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
async-timeout==3.0.1
atomicwrites==1.1.5
attrs==18.1.0
autobahn==19.7.1
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.1
billiard==3.3.0.23
binaryornot==0.4.4
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
celery==3.1.26.post2
certifi==2019.3.9
cffi==1.12.2
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
click==6.7
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.6.11
conda-build==3.17.8
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cookiecutter==1.6.0
coverage==4.5.1
cryptography==2.7
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
daphne==2.3.0
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
distributed==1.26.0
dj-database-url==0.5.0
Django==2.2.2
django-ajax-auth==0.1.1
django-background-tasks==1.2.5
django-bootstrap-modal-forms==1.4.2
django-bootstrap3==11.1.0
django-braces==1.13.0
django-celery==3.3.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-jquery==3.1.0
django-model-utils==3.1.2
django-star-ratings==0.8.0
django-webtest==1.9.3
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fancycompleter==0.8
fastcache==1.0.2
filelock==3.0.10
first==2.0.1
flake8==3.5.0
Flask==1.0.2
future==0.17.1
gevent==1.4.0
glob2==0.6
greenlet==0.4.15
gunicorn==19.9.0
h5py==2.9.0
heapdict==1.0.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
hypothesis==1.16.0
idna==2.6
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
menuinst==1.4.16
misaka==2.1.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
model-mommy==1.6.0
more-itertools==4.3.0
mpmath==1.1.0
mqtt==0.0.1
msgpack==0.5.6
multidict==4.5.2
multipledispatch==0.6.0
mysqlclient==1.4.5
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.1
packaging==19.0
paho-mqtt==1.5.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.0.1
pathlib2==2.3.3
patsy==0.5.1
pbr==4.2.0
pdbpp==0.9.2
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
pip-tools==2.0.2
pkginfo==1.5.0.1
pluggy==0.7.1
ply==3.11
poyo==0.4.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.1
psycopg2==2.8.4
psycopg2-binary==2.7.7
py==1.5.4
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==3.7.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-cov==2.5.1
pytest-django==3.3.3
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.5
PyWavelets==1.0.2
pywin32==223
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
redis==2.10.3
requests==2.18.4
requests-oauthlib==1.2.0
rope==0.12.0
ruamel-yaml==0.15.46
schedule==0.6.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
selenium==3.14.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
slugify==0.0.1
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
swapper==1.1.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
waitress==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.2
WebTest==2.0.30
Werkzeug==0.14.1
whichcraft==0.6.0
whitenoise==4.1.2
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wmctrl==0.3
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
yarl==1.3.0
zict==0.1.4
zipp==0.3.3
zope.interface==4.6.0
